Refactor to use router
