+++ src/app/app.component.css
@@ -1,304 +1,26 @@
-@import url('https://fonts.googleapis.com/css?family=Roboto|Roboto+Condensed|Roboto+Mono|Roboto+Slab&display=swap');
-
-* {
-  font-family: 'Roboto Condensed', sans-serif;
-  font-size: 18px;
-  -webkit-transition: all 0.5s ease;
-  -moz-transition: all 0.5s ease;
-  -o-transition: all 0.5s ease;
-  transition: all 0.5s ease;
-  margin: 0px;
-  padding: 0px;
-}
-
-/* HTML5 display-role reset for older browsers */
-article,
-aside,
-details,
-figcaption,
-figure,
-footer,
-header,
-hgroup,
-menu,
-nav,
-section {
-  display: block;
-}
-body {
-  line-height: 1;
-  background-color: #efefef;
-  height: 100%;
-  overflow-x: hidden;
-}
-ol,
-ul {
-  list-style: none;
-}
-blockquote,
-q {
-  quotes: none;
-}
-h1 {
-  font-family: 'nunitoregular';
-  color: #707070;
-}
-.float-right {
-  float: right;
-}
-.opacity-60 {
-  opacity: 0.6;
-}
-.letter-spacing {
-  letter-spacing: 1px;
-}
-.text-transform-uppercase {
-  text-transform: uppercase;
-}
-.font-size-20 {
-  font-size: 20px !important;
-}
-.center {
-  top: 50%;
-  left: 55%;
-  position: absolute;
-  -webkit-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-  width: 100%;
-}
-
 header {
-  background-image: url('../assets/images/strip_top.png');
+  background-image: url('/assets/images/strip_top.png');
   background-repeat: no-repeat;
   background-size: 100%;
   height: 130px;
 }
+
 header h1 {
   /* There's a problem with this: if the page is too narrow, this ends up being white on white. */
   padding: 16px 0 0 64px;
+  margin: 0;
 }
-.color1 {
-  color: #707070;
-}
-.color2 {
-  color: #525151;
-}
-.color3 {
-  color: #66d5e3;
-}
-.color4 {
-  color: #716f6f;
-}
-.color7 {
-  color: #a5a3a3;
-}
-.color8 {
-  color: #ffffff !important;
-}
-.color11 {
-  color: #9d9d9d;
-}
-.header {
-  font-family: Roboto;
-  font-weight: normal;
-  font-size: 45px;
-  text-align: left;
-  color: #707070;
-}
-.header2 {
-  font-family: Roboto;
-  font-weight: 900;
-  font-size: 65px;
-  letter-spacing: 0.03em;
-  line-height: 120px;
-  text-align: left;
-  color: #525151;
-}
-.subText {
-  height: 63px;
-  background: transparent;
-}
-/* ::placeholder {
-  / * Chrome, Firefox, Opera, Safari 10.1+ * /
-  font-family: Roboto;
-  font-weight: 300;
-  font-size: 21px;
-  text-align: left;
-  padding-left: 10px;
-  color: #707070;
-} */
-.containingDIV {
-  display: flex;
-}
-.footer {
+
+footer {
   position: fixed;
   left: 0;
   bottom: 0;
   width: 100%;
   color: white;
   text-align: center;
-  background-image: url('../assets/images/strip_bottom.png');
+  background-image: url('/assets/images/strip_bottom.png');
   height: 120px;
   background-size: 100%;
   background-repeat: no-repeat;
   z-index: -1;
-}
-.mi-margin-10 {
-  margin: 10px;
-}
-.mi-margin-5 {
-  margin: 5%;
-}
-.mi-margin-left-10 {
-  margin-left: 10px;
-}
-.matching {
-  font-family: Roboto;
-  font-weight: 900;
-  font-size: 25px;
-  text-align: left;
-  margin-top: 12px;
-}
-
-.filters {
-  width: 160px;
-  border-radius: 5px;
-  text-align: center;
-  background: #fff;
-  border: 1px solid #707070;
-  padding: 5px;
-}
-
-.cards {
-  border-radius: 3px;
-  border: 1px solid #cbcbcb;
-  margin-bottom: 10px;
-  background-color: white;
-}
-.cardText2 {
-  font-family: Roboto;
-  font-weight: 500;
-  font-size: 21px;
-  text-align: left;
-  color: #525151;
-}
-
-/**10***/
-
-.padding-top-10 {
-  padding-top: 10px;
-}
-
-.padding-top-15 {
-  padding-top: 15px;
-}
-
-.padding-bottom-20 {
-  padding-bottom: 20px;
-}
-.padding-left-20 {
-  padding-left: 20px;
-}
-.padding-top-30 {
-  padding-top: 30px;
-}
-.col100 {
-  width: 100%;
-}
-* {
-  box-sizing: border-box;
-}
-
-/* Clear floats after the columns */
-.row:after {
-  content: '';
-  display: table;
-  clear: both;
-}
-
-.dropdown-content {
-  display: none;
-  position: absolute;
-  background-color: #ffffff;
-  min-width: 300px;
-  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
-  padding: 12px 16px;
-  z-index: 1;
-  margin: 2px 0;
-  height: 300px;
-  text-align: left;
-}
-.dropdown-content ul {
-  margin: 15px 0;
-}
-.filters:hover .dropdown-content {
-  display: block;
-}
-
-.selectedPage {
-  color: #66d5e3;
-}
-::-webkit-scrollbar {
-  width: 10px;
-}
-
-/* Track */
-::-webkit-scrollbar-track {
-  box-shadow: inset 0 0 2px #a5a3a3;
-  border-radius: 5px;
-}
-
-/* Handle */
-::-webkit-scrollbar-thumb {
-  background: #66d5e3;
-  border-radius: 5px;
-}
-
-.pointer:hover {
-  color: #000;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.facilityList {
-  overflow-y: auto;
-  max-height: 100px;
-  color: #898686;
-  padding: 10px;
-}
-
-.facilityList .phone,
-.facilityList .email {
-  font-size: 80%;
-  color: #0079e0;
-}
-
-.facility {
-  margin-bottom: 15px;
-}
-
-.phoneIcon,
-.emailIcon {
-  display: inline-block;
-  width: 20px;
-  height: 20px;
-  background-repeat: no-repeat;
-  text-indent: 100%;
-  white-space: nowrap;
-  overflow: hidden;
-}
-
-.phoneIcon {
-  background-image: url(/assets/images/phone.svg);
-}
-.emailIcon {
-  background-image: url(/assets/images/email.svg);
-}
-
-.view-record-text {
-  color: #0079e0;
-}
-
-.view-record-text:hover {
-  color: #ff00b3;
 }
