+++ src/app/app.component.html
@@ -1,13 +1,5 @@
-<custom-spinner
-  class="topVisible"
-  *ngIf="showOverlay"
-  [value]="loadingPercentage"
-  [mode]="loadingMode"
-  [loadingText]="loadingText"
-></custom-spinner>
-
 <header>
-  <h1 (click)="backToHomePage()">
+  <h1 (click)="navigateToSearch()">
     <svg xmlns="http://www.w3.org/2000/svg" width="33.5" height="40.926" viewBox="0 0 33.5 40.926">
       <path
         id="iconfinder_laboratory_microscope_research_science__5310256"
@@ -31,56 +23,5 @@
     </svg>
   </h1>
 </header>
-<div class="row">
-  <!-- Home page-->
-  <div *ngIf="!searchPage">
-    <h1 class="color1 header">Welcome {{ (patient | async)?.getUsualName() }},</h1>
-    <h1 class="color2 header2">Let’s find some clinical trials</h1>
-    <h1 class="color1 subText">
-      We found your <span class="view-record-text" (click)="showRecord()">record</span>. Just fill in some preferences
-      and we’ll do the rest
-    </h1>
-    <app-search-fields (searchClicked)="searchClinicalTrials($event)"></app-search-fields>
-    <!--Patient Records-->
-    <app-record-data
-      [(displayOn)]="this.records"
-      [bundleResources]="this.bundleResources"
-      [patient]="this.patient"
-      max-height="90vh"
-      height="90%"
-      vertical-align="top"
-      position="absolute"
-      overflow="scroll"
-      overflow-y="scroll"
-    >
-    </app-record-data>
-  </div>
-  <!-- Cards page-->
-  <div class="mi-margin-5" style="margin-top: 9%;" *ngIf="!searchtable">
-    <app-results [searchResults]="searchResults" [searchParameters]="searchReqObject"></app-results>
-  </div>
-  <app-result-details
-    *ngIf="!detailsPage"
-    [clinicalTrial]="detailedTrial"
-    [trialSaved]="savedClinicalTrialsNctIds.has(detailedTrial.nctId)"
-    [reqs]="searchReqObject"
-    (trialSaveChanged)="setTrialSaved(detailedTrial, $event)"
-  >
-    <div class="mi-margin-10 col100 containingDIV pointer" (click)="backToSearch()">
-      <svg xmlns="http://www.w3.org/2000/svg" width="40.641" height="40.641" viewBox="0 0 40.641 40.641">
-        <g id="arrow_upward-24px" transform="translate(0 40.641) rotate(-90)">
-          <path id="Path_20" data-name="Path 20" d="M0,0H40.641V40.641H0Z" fill="none" />
-          <path
-            id="Path_21"
-            data-name="Path 21"
-            d="M4,17.547l2.388,2.388,9.466-9.449V31.094H19.24V10.486l9.449,9.466,2.4-2.4L17.547,4Z"
-            transform="translate(2.773 2.773)"
-            fill="#efefef"
-          />
-        </g></svg><span class="color8" style="margin-top: 10px;"> Back to results</span>
-    </div>
-  </app-result-details>
-  <div class="row footer" *ngIf="!searchPage">
-    <div class="col-sm-4"></div>
-  </div>
-</div>
+<router-outlet></router-outlet>
+<footer></footer>
