+++ src/app/bundle.ts
@@ -1,14 +1,22 @@
 import { Bundle, BundleEntry, Resource } from './fhir-types';
+import { TrialQuery } from './services/search-results.service';
 
 /**
  * Values that can be placed into parameters
  */
 type Stringable = string | number | boolean | null;
 
+export interface ParameterResource extends Resource {
+  resourceType: 'Parameters';
+  parameter?: { name: string; valueString: Stringable }[];
+}
+
+export type PatientBundle = Bundle;
+
 /**
  * Create collection bundle from parameters and entries
  */
-export function createPatientBundle(parameters: { [key: string]: Stringable }, entries: BundleEntry[]): PatientBundle {
+export function createPatientBundle(parameters: TrialQuery, entries: BundleEntry[]): PatientBundle {
   const paramResource: ParameterResource = { resourceType: 'Parameters', id: '0', parameter: [] };
   for (const p in parameters) {
     // Ignore null values, they indicate a parameter that should be ignored.
@@ -26,10 +34,3 @@ export function createPatientBundle(parameters: { [key: string]: Stringable }, e
   });
   return patientBundle;
 }
-
-export interface ParameterResource extends Resource {
-  resourceType: 'Parameters';
-  parameter?: { name: string; valueString: Stringable }[];
-}
-
-export type PatientBundle = Bundle;
